requirements.txt
streamlit~=1.2.0
pillow~=8.4.0
numpy~=1.21.4
matplotlib~=3.5.0
torch~=1.10.0
torchvision~=0.11.1